feat(fe): remove deprecated methods sdk from app
